team crud auth, unit tests

+++ app/Http/Controllers/TeamController.php
@@ -5,9 +5,9 @@ namespace App\Http\Controllers;
 use App\Http\Requests\TeamRequest;
 use App\Services\TeamService\TeamServiceInterface;
 use Illuminate\Http\Response;
-use Illuminate\Support\Facades\Auth;
 
 use App\Models\User;
+use Illuminate\Support\Facades\Auth;
 
 class TeamController extends Controller
 {
@@ -27,10 +27,9 @@ class TeamController extends Controller
 
     public function create(TeamRequest $request)
     {
-//        if (!$user = auth()->user()){
-//            return $this->apiResponse(Response::HTTP_NOT_FOUND, null, 'User not authenticated');
-//        }
-        $user = User::first();
+        if (!$user = auth()->user()){
+            return $this->apiResponse(Response::HTTP_NOT_FOUND, null, 'User not authenticated');
+        }
 
         $content = $this->teamService->create($user->id, $request->all());
 
@@ -52,10 +51,9 @@ class TeamController extends Controller
 
     public function update(TeamRequest $request, $teamId)
     {
-//        if (!$user = auth()->user()){
-//            return $this->apiResponse(Response::HTTP_NOT_FOUND, null, 'User not authenticated');
-//        }
-        $user = User::first();
+        if (!$user = auth()->user()){
+            return $this->apiResponse(Response::HTTP_NOT_FOUND, null, 'User not authenticated');
+        }
 
         $content = $this->teamService->update($user->id, $teamId, $request->all());
 
@@ -64,10 +62,9 @@ class TeamController extends Controller
 
     public function delete($teamId)
     {
-//        if (!$user = auth()->user()){
-//            return $this->apiResponse(Response::HTTP_NOT_FOUND, null, 'User not authenticated');
-//        }
-        $user = User::first();
+        if (!$user = auth()->user()){
+            return $this->apiResponse(Response::HTTP_NOT_FOUND, null, 'User not authenticated');
+        }
 
         if (!$this->teamService->delete($teamId)){
             return $this->apiResponse(Response::HTTP_NOT_FOUND, null, 'Team not found');
